Create a basic git-authors view for Atom
Addresses
---------
Need to know all authors for a file.

Why
---
I've chosen the simplest solution for now: you view the list of authors and
their %age contribution in a modal panel. This proves out the solution because
git needs to be called through a BufferedProcess to get the output of blame.
I use the `--line-porcelain` option so that it can be processed easily, and
then filter for the lines containing author information.

Next up: change the display to be non-modal; hopefully as a display that rises
up from the status bar.

What
----